members_automation/requirements.txt
click==8.1.3

joblib==1.1.0
pandas==1.4.2
numpy==1.22.3

gspread==5.7.1
selenium==4.9.1
webdriver-manager==3.8.6

PyYAML==6.0
python-dotenv==1.0.0
python-box==7.0.1
chromedriver-binary